requirements.txt
psutil==5.8.0
requests==2.25.1
python-dotenv==1.2.1